src/main/java/com/anyservice/service/user/PasswordService.java
package com.anyservice.service.user;

import com.anyservice.service.api.IPasswordService;
import org.mindrot.jbcrypt.BCrypt;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Service;

@Service
public class PasswordService implements IPasswordService {

    @Value("${password.rounds}")
    private int logRounds;

    public String hash(String password) {
        return BCrypt.hashpw(password, BCrypt.gensalt(logRounds));
    }

    public boolean verifyHash(String password, String hash) {
        return BCrypt.checkpw(password, hash);
    }

}
